+++ backend/src/data/models/theme.ts
@@ -17,7 +17,7 @@ export const Theme = model<ThemeDocument>('theme', new Schema({
     default: generateInvite,
     unique: [true, 'Code should be unique'],
     dropDups: true,
-    validate: [/(?<!discord|accord|default)$/, 'This code is reserved'],
+    validate: [/(?<!discord|anichat|default)$/, 'This code is reserved'],
     maxlength: [32, 'Code is too long'],
   },
   createdAt: { type: Date, default: new Date() },