+++ frontend/src/components/navigation/navbar.tsx
@@ -8,7 +8,7 @@ const Navbar: React.FunctionComponent = () => {
   return (
     <nav className="flex items-center justify-between h-15 p-4 px-8">
       <a className="logo">
-        <span className="font-bold heading">acrd</span>
+        <span className="font-bold heading">anichat</span>
         <span className="normal">.</span>
         <span className="muted secondary">app</span>
       </a>